+++ mex/editor/state.py
@@ -36,6 +36,11 @@ class State(rx.State):
             raw_path="/",
         ),
         NavItem(
+            title="layout.nav_bar.advanced_search_navitem",
+            route_ids=["/advanced-search"],
+            raw_path="/advanced-search/?page=1",
+        ),
+        NavItem(
             title="layout.nav_bar.create_navitem",
             route_ids=["/create", "/create/[draft_id]"],
             raw_path="/create",
@@ -69,9 +74,13 @@ class State(rx.State):
         return [self._translate_nav_item(item) for item in self._nav_items]
 
     @rx.event
-    def set_is_unsaved_changes_dialog_open(self, value: bool) -> None:  # noqa: FBT001
-        """Set the unsaved changes dialog open state."""
-        self.is_unsaved_changes_dialog_open = value
+    def set_is_unsaved_changes_dialog_open(self, is_open: bool) -> None:  # noqa: FBT001
+        """Set the state of the unsaved changes dialog.
+
+        Args:
+            is_open: Whether the dialog should be open or not.
+        """
+        self.is_unsaved_changes_dialog_open = is_open
 
     @rx.event
     def change_locale(self, locale: str) -> None: