+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- New page for advanced searching, including complex reference filter
+- Translations for entity types
 - Button on edit page to reset or delete rules
 - Warn about unsaved changes when logging out
 - pytest-retry since tests often failing on github (we belief due to low system/cpu performance in CI)
@@ -17,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - upgrade reflex to 0.8.15
 - use pydantic v2 BaseModel instead of rx.Base
+- other rerun lib for pytest
 - updated mex-common to 1.16.0
 - updated template to https://github.com/robert-koch-institut/mex-template/commit/81bc5b