+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - change primary source of rules from mex to mex-editor
+- moved pytest rerun config from toml to makefile
 
 ### Deprecated
 
@@ -21,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - fixed width of search and advanced-search result lists
 - fixed redirect target after login screen interception
-- fixed duplicate settings logging on startup
+- fixed settings logging on startup and in tests
 
 ### Security