+++ CHANGELOG.md
@@ -17,10 +17,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+### Security
+
+## [1.5.1] - 2026-02-25
+
+### Fixed
+
 - fixed width of search and advanced-search result lists
 - fixed redirect target after login screen interception
-
-### Security
 
 ## [1.5.0] - 2026-02-24